Migrate microsoft_drive to new format.
