+++ front/lib/actions/mcp_internal_actions/constants.ts
@@ -31,6 +31,7 @@ import { HUBSPOT_SERVER } from "@app/lib/api/actions/servers/hubspot/metadata";
 import { IMAGE_GENERATION_SERVER } from "@app/lib/api/actions/servers/image_generation/metadata";
 import { INCLUDE_DATA_SERVER } from "@app/lib/api/actions/servers/include_data/metadata";
 import { JIT_TESTING_SERVER } from "@app/lib/api/actions/servers/jit_testing/metadata";
+import { MICROSOFT_DRIVE_SERVER } from "@app/lib/api/actions/servers/microsoft_drive/metadata";
 import { MICROSOFT_EXCEL_SERVER } from "@app/lib/api/actions/servers/microsoft_excel/metadata";
 import { MICROSOFT_TEAMS_SERVER } from "@app/lib/api/actions/servers/microsoft_teams/metadata";
 import { MISSING_ACTION_CATCHER_SERVER } from "@app/lib/api/actions/servers/missing_action_catcher/metadata";
@@ -794,19 +795,19 @@ export const INTERNAL_MCP_SERVERS = {
     tools_arguments_requiring_approval: undefined,
     tools_retry_policies: undefined,
     timeoutMs: undefined,
-    serverInfo: {
-      name: "microsoft_drive",
-      version: "1.0.0",
-      description: "Tools for managing Microsoft files.",
-      authorization: {
-        provider: "microsoft_tools" as const,
-        supported_use_cases: ["personal_actions"] as const,
-        scope:
-          "User.Read Files.ReadWrite.All Sites.Read.All ExternalItem.Read.All offline_access" as const,
+    metadata: {
+      ...MICROSOFT_DRIVE_SERVER,
+      serverInfo: {
+        ...MICROSOFT_DRIVE_SERVER.serverInfo,
+        authorization: {
+          provider: "microsoft_tools" as const,
+          supported_use_cases: ["personal_actions"] as const,
+          scope:
+            "User.Read Files.ReadWrite.All Sites.Read.All ExternalItem.Read.All offline_access" as const,
+        },
+        documentationUrl:
+          "https://docs.dust.tt/docs/microsoft-drive-tool-setup",
       },
-      icon: "MicrosoftLogo",
-      documentationUrl: "https://docs.dust.tt/docs/microsoft-drive-tool-setup",
-      instructions: null,
     },
   },
   microsoft_teams: {
